+++ logicx_ishop/patches/add_item_ishop_tab.py
@@ -1,0 +1,5 @@
+from logicx_ishop.custom.item import setup_item_custom_fields
+
+
+def execute():
+	setup_item_custom_fields()